requirements.txt
python-binance>=1.0.19
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
flask>=3.0.0
flask-cors>=4.0.0
